Make a pleasing diagnostic when core modules are missing
This may be somewhat passive aggresive, but I prefer to know
when a core module is missing from my Perl, rather than immediately
going to CPAN to download it.  The primary argument against removing
core prereqs is that the existing diatnostic "Env Not found" or
whatever is confusing to those that do not recognize a core module.

This is my compromise.  Detect when the core module is missing and
barf with a helpful error message directing them to manually fix
up their mess.